sessions/dbt-100/101/requirements.txt
asn1crypto==1.4.0
azure-common==1.1.27
azure-core==1.13.0
azure-storage-blob==12.8.1
boto3==1.17.69
botocore==1.20.69
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.7
idna==2.10
isodate==0.6.0
jmespath==0.10.0
msrest==0.6.21
numpy==1.18.2
oauthlib==3.1.0
oscrypto==1.2.1
pycparser==2.20
pycryptodomex==3.10.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
snowflake-connector-python==2.4.3
urllib3==1.26.4
